requirements.txt
streamlit
langchain
langchain_chroma
langchain_huggingface
langchain_community
langchain_google_genai
python-dotenv